fix(table): repair what the first pass broke, and eight defects it uncovered
A line-by-line audit of the previous two commits. Three of these are regressions
this branch introduced; the rest were already there and sit in the same paths.

REGRESSIONS FROM THIS BRANCH

1. THE SPOTLIGHT DIED. Generalising the hero override to `.seat--in-hand
   { filter: none !important }` stopped live villains being faded and took five
   other things with it, because that declaration lands on every seat in a live
   hand: the acting-player brightness (.seat-wrapper--spot), the gold
   stack-change glow, avatar hover, the bust-art drop shadow, and it out-ranked
   .seat--folded during the window where lastAction is 'fold' but status has not
   flipped - putting a bright frame on every fold. Reverted to hero-only. The
   villain fade is fixed one level up in seatDimmed, which no longer dims any
   seat holding a live hand, so there is nothing left here to override.

2. THE BUTTON COULD STAND STILL. getNextSeat over a one-seat roster returns that
   seat from both branches, so a single button-eligible player already on the
   button kept it and the same two players posted the small and big blind twice
   running - reachable whenever several players arrive at once around one
   incumbent. The button is now forced across the full roster in that case.
   Heads-up is deliberately exempt: with two players the button IS the small
   blind, so forcing it would put the newcomer in the small blind, the hold-out
   would refuse them, and the table would never deal again.

3. BUTTON ELIGIBILITY LEAKED INTO TOURNAMENTS. Nobody "sits down" in a
   tournament: the seating sweep places late registrants and TableBalancer moves
   players between tables deliberately, positioning them relative to the big
   blind. Filtering them out of the rotation silently overrode that placement,
   and a freshly balanced table is mostly players the set has never seen. Now
   cash-only.

FOUND IN THE SAME PATHS

4. A TABLE SETTING COULD SWITCH OFF A HOUSE RULE. registerWaitForBB was gated on
   wait_for_big_blind, and that set is what the hold-outs read. A host who
   turned it off skipped registration, skipped the hold-out, and had brand-new
   players dealt straight into the small blind.

5. POST /post-bb BECAME A WAY TO PAY TO BREAK THAT RULE. Once entry was free the
   only players still waiting were the two being held out, so the endpoint
   stopped being a shortcut and became the one way to pay a live big blind to be
   dealt into the small blind. It now refuses that seat.

6. RUN IT TWICE SOLD FABRICATED CARDS. On a RIT hand communityCards holds only
   the shared pre-all-in prefix, so the rabbit-hunt gate read board length 0,
   decided the hand ended pre-flop, and charged five diamonds for five cards -
   off a deck RIT had already burned two or three run-outs from. Not "what would
   have come", noise. RIT hands are no longer offered a hunt.

7. A PAID REVEAL COULD BE TORN DOWN MID-ANIMATION. Awaiting 500ms per card
   before the first appeared meant a pre-flop fold took 2.5s to draw; the next
   hand starting inside that window unmounted the panel and the player had paid
   for cards they never saw. All cards are set at once, CSS staggers them.

8. THE BBJ CELEBRATION WAS CUT SHORT. The client never passed bbjHit to
   handCompletionHoldMs, so it held ~7.9s against the server's 9s and wiped the
   board a second into the Bad Beat Jackpot celebration. Animation speed is now
   clamped too: below 1x it cleared the winner name before the pot-win float it
   describes had finished, above it, ~12s of stale board survived a dropped
   HAND_STARTED where the old hardcoded 3000ms risked one second.

9. THE VIP COUNTER WAS COMPUTED AND THROWN AWAY. vip_remaining came all the way
   from the RPC to TablePage and was dropped one line from the UI, so the button
   said FREE on the 101st hunt and silently took five diamonds. Now shown, and
   the label stops saying FREE when the pool is spent. VIPPage still advertised
   "Unlimited"; it says 100 / month.

10. SMALLER: the offer history was trimmed on `handNumber < handCount - 1`, but
    hand numbers are global and jump, so it kept one entry instead of two and
    halved the window for a late click - now trimmed by insertion order. The
    capture catch was silent, and was the blind spot that would have hidden #6.
    A hung VIP lookup could disable the button forever. The handler answered
    every server-side fault with "Invalid request body". Every billing refusal
    read "Could Not Complete Purchase". diamonds_remaining serialised as a
    string. The horse auto-cashout predicted the next button with a fourth,
    independent walk that could disagree with the other three. Two applied CSS
    classes were defined nowhere. One dead import.

The "Post BB To Enter" overlay is now a notice: the wait it offered to skip is
one hand and free, and the engine refuses the only call it could still make.

fn_consume_rabbit_hunt reapplied to production for the numeric cast.
Three source-text tests pinned shapes changed here and are updated beside them.
Client 5524 tests green, server 1572 green, both typechecks clean.

+++ supabase/migrations/20260825_fn_consume_rabbit_hunt.sql
@@ -130,7 +130,9 @@ BEGIN
       'success', true,
       'source', 'diamonds',
       'diamonds_spent', v_cost,
-      'diamonds_remaining', (v_deduct->>'balance')
+      -- ->> yields TEXT, so this used to serialise as "42" rather than 42 while
+      -- the API type promised a number. ::numeric makes the wire format match.
+      'diamonds_remaining', (v_deduct->>'balance')::numeric
     );
   END IF;
